+++ test/unit/fuzz-tools.test.js
@@ -103,6 +103,7 @@ import * as cardioV101 from '../../lib/cardio-v101.js';
 import * as cardioV102 from '../../lib/cardio-v102.js';
 import * as cvriskV103 from '../../lib/cvrisk-v103.js';
 import * as cardioV104 from '../../lib/cardio-v104.js';
+import * as vascularV105 from '../../lib/vascular-v105.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -150,6 +151,7 @@ const MODULES = {
   'cardio-v102.js': cardioV102,
   'cvrisk-v103.js': cvriskV103,
   'cardio-v104.js': cardioV104,
+  'vascular-v105.js': vascularV105,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];